Fix build break

+++ src/buffer_handler_wayland.c
@@ -1072,7 +1072,7 @@ static inline widget_fb_t raw_open_file(const char *filename)
 	buffer->state = WIDGET_FB_STATE_CREATED;
 	buffer->type = WIDGET_FB_TYPE_FILE;
 	buffer->refcnt = 0;
-	buffer->info = (void *)off;
+	buffer->info = (void *)((unsigned long)off);
 
 	ret = read(fd, buffer->data, off);
 	if (ret < 0) {
